# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workspace API",
    "version": "2.2.0",
    "description": "REST API for managing files within a Linux workspace directory. All requests must be authenticated using a Bearer token provided in the Authorization header."
  },
  "servers": [
    {
      "url": "https://workspace.foliole.eu",
      "description": "Workspace Endpoint"
    }
  ],
  "paths": {
    "/files": {
      "get": {
        "summary": "List all files recursively",
        "operationId": "listFiles",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of files returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: Invalid token"
          }
        }
      },
      "post": {
        "summary": "Create a new file",
        "operationId": "createFile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "path"
                ],
                "properties": {
                  "path": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: Invalid token"
          },
          "409": {
            "description": "Conflict: File already exists"
          }
        }
      },
      "delete": {
        "summary": "Delete a file",
        "operationId": "deleteFile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "path"
                ],
                "properties": {
                  "path": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: Invalid token"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/files/{path}": {
      "get": {
        "summary": "Read a file",
        "operationId": "readFile",
        "security": [
          {
            "GPTSecret": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File content returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: Invalid token"
          },
          "404": {
            "description": "File not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing file",
        "operationId": "updateFile",
        "security": [
          {
            "GPTSecret": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: Invalid token"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/files/rename": {
      "post": {
        "summary": "Rename a file",
        "operationId": "renameFile",
        "security": [
          {
            "GPTSecret": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "old_path",
                  "new_path"
                ],
                "properties": {
                  "old_path": {
                    "type": "string"
                  },
                  "new_path": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File renamed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "old_path": {
                      "type": "string"
                    },
                    "new_path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: Invalid token"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/files/move": {
      "post": {
        "summary": "Move a file",
        "operationId": "moveFile",
        "security": [
          {
            "GPTSecret": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "src_path",
                  "dest_path"
                ],
                "properties": {
                  "src_path": {
                    "type": "string"
                  },
                  "dest_path": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File moved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "src_path": {
                      "type": "string"
                    },
                    "dest_path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: Invalid token"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/commit": {
      "post": {
        "summary": "Commit changes to Git repository",
        "operationId": "commitChanges",
        "security": [
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "Commit message for Git"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Changes committed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "commit_message": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": { "description": "Forbidden: Invalid token" },
          "500": { "description": "Git commit failed" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {}
  }
}